+++ client/Sheepshead/SharedSheep/Blind/IBlind.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using SharedSheep.Card;
+
+namespace SharedSheep.Blind
+{
+    interface IBlind
+    {
+        List<ICard> BlindCards { get; set; }
+        void AddCard(ICard card);
+        ICard SwapCard(int index, ICard card);
+        int BlindPoints();
+    }
+}
