+++ client/Sheepshead/SharedSheep/Common/Enums.cs
