+++ client/Sheepshead/SharedSheep/Hand/IHand.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using SharedSheep.Card;
+
+namespace SharedSheep.Hand
+{
+    public interface IHand
+    {
+        List<ICard> Cards { get; set; }
+
+        int GetNumOfRemainingCards();
+        ICard GetCard(int index);
+        int NumOfThisSuit(Suit suit);
+    }
+}
