+++ client/Sheepshead/SharedSheep/SharedSheep.csproj
@@ -1,0 +1,16 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <Folder Include="Card\" />
+    <Folder Include="Common\" />
+    <Folder Include="Deck\" />
+    <Folder Include="Hand\" />
+    <Folder Include="Blind\" />
+    <Folder Include="Player\" />
+  </ItemGroup>
+
+</Project>
