+++ client/Sheepshead/SharedSheep/Trick/ITrick.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using SharedSheep.Player;
+using SharedSheep.Card;
+
+namespace SharedSheep.Trick
+{
+    interface ITrick
+    {
+        List<(IPlayer, ICard)> TrickCards { get; }
+
+        IPlayer TheWinnerPlayer();
+        ICard TheWinnerCard();
+        ICard LeadingCard();
+
+
+    }
+}
