+++ client/Sheepshead/Sheepshead.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.27130.2027
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SharedSheep", "SharedSheep\SharedSheep.csproj", "{2B7D5964-2091-4433-AD98-80E0776AEB34}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SharedSheepTest", "SharedSheepTest\SharedSheepTest.csproj", "{36B51193-74F2-4881-B050-5C7DBD3182B8}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{2B7D5964-2091-4433-AD98-80E0776AEB34}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2B7D5964-2091-4433-AD98-80E0776AEB34}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2B7D5964-2091-4433-AD98-80E0776AEB34}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2B7D5964-2091-4433-AD98-80E0776AEB34}.Release|Any CPU.Build.0 = Release|Any CPU
+		{36B51193-74F2-4881-B050-5C7DBD3182B8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{36B51193-74F2-4881-B050-5C7DBD3182B8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{36B51193-74F2-4881-B050-5C7DBD3182B8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{36B51193-74F2-4881-B050-5C7DBD3182B8}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {E5C08CE7-1C0D-4686-BC9C-8D49B5677480}
+	EndGlobalSection
+EndGlobal
