--- client/sheepshead/sheepshead.sln
@@ -1,33 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.27130.2026
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "sheepshead", "sheepshead\sheepshead.csproj", "{CD8E48DF-1887-4AF4-9BC6-D858E267170B}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "sheepsheadTests", "sheepsheadTests\sheepsheadTests.csproj", "{ABF43EB9-AABB-4FDA-8D51-5CA6CC053416}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{CD8E48DF-1887-4AF4-9BC6-D858E267170B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CD8E48DF-1887-4AF4-9BC6-D858E267170B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CD8E48DF-1887-4AF4-9BC6-D858E267170B}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
-		{CD8E48DF-1887-4AF4-9BC6-D858E267170B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CD8E48DF-1887-4AF4-9BC6-D858E267170B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{CD8E48DF-1887-4AF4-9BC6-D858E267170B}.Release|Any CPU.Deploy.0 = Release|Any CPU
-		{ABF43EB9-AABB-4FDA-8D51-5CA6CC053416}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{ABF43EB9-AABB-4FDA-8D51-5CA6CC053416}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{ABF43EB9-AABB-4FDA-8D51-5CA6CC053416}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{ABF43EB9-AABB-4FDA-8D51-5CA6CC053416}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {EB639715-D45F-46A5-9258-D264E13EF2BE}
-	EndGlobalSection
-EndGlobal
