--- client/sheepshead/sheepshead/Properties/AndroidManifest.xml
@@ -1,9 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android" 
-          package="sheepshead.sheepshead" 
-          android:versionCode="1" 
-          android:versionName="1.0">
-  <uses-sdk android:minSdkVersion="21" />
-  <application android:allowBackup="true" android:label="@string/app_name">
-  </application>
-</manifest>
