--- client/sheepshead/sheepshead/Resources/models/interfaces/IHand.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace sheepshead.Resources.models.interfaces
-{
-    public interface IHand
-    {
-        List<ICard> Cards { get; set; }
-
-        int GetNumOfRemainingCards();
-        ICard GetCard(int index);
-        int NumOfThisSuit(Suit suit);
-    }
-}
