--- client/sheepshead/sheepshead/Resources/models/interfaces/IPlayer.cs
@@ -1,15 +1,0 @@
-﻿using System;
-namespace sheepshead.Resources.models.interfaces
-{
-    public interface IPlayer
-    {
-        IHand Hand { get; set; }
-        String Name { get; }
-
-        Boolean IsPartner();
-        ICard PlayCard();
-        Boolean WantPick();
-        // IBlind Pick(IBlind blind);
-
-    }
-}
