--- client/sheepshead/sheepshead/Resources/values/Strings.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-  <string name="app_name">sheepshead</string>
-</resources>
