--- client/sheepshead/sheepshead/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="System.ValueTuple" version="4.4.0" targetFramework="monoandroid71" />
-</packages>
