--- client/sheepshead/sheepsheadTests/Resources/models/TrickTests.cs
@@ -1,21 +1,0 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using sheepshead.Resources.models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-using System.Threading.Tasks;
-
-namespace sheepshead.Resources.models.Tests
-{
-    [TestClass()]
-    public class TrickTests
-    {
-        [TestMethod()]
-        public void TheWinnerCardTest()
-        {
-            Assert.Fail();
-        }
-    }
-}
