--- client/sheepshead/sheepsheadTests/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="MSTest.TestAdapter" version="1.2.0" targetFramework="net45" />
-  <package id="MSTest.TestFramework" version="1.2.0" targetFramework="net45" />
-</packages>
